Add support for LLVM Flang
LLVM Flang is the new Fortran frontend from LLVM. It is still
work-in-progress, but there is enough support for Fortran 95/03 and
OpenMP to compile and run SNAP. To enable this, a new target is defined
in Makefile (`fsnap`) that corresponds to SNAP being built with LLVM
Flang.

LLVM Flang is under active development and it is assumed that you will
be building it from sources in order to compile SNAP. For this reason,
no assumptions are being made with respect to where to locate it (this
is in contrast to regular system compilers which are usually available
through $PATH). Instead, you will have point Makefile to it through the
`LLVM_FLANG_DIR` variable.

You will also have to specify where to find PGMATH, which LLVM Flang
currently depends on for Math routines. You will specify this with
`PGM_DIR`.

Last, but not least, the LLVM Flang compiler driver is currently called
`flang-new`. This is likely to change in the near future at which point
the build script will have to be updated accordingly.

Co-authored-by: Mats Petersson <mats.petersson@arm.com>